fix: Exception when id_token not set.
django-oauth-toolkit does not support openid for the resource owner
credentials flow. The response will still have openid in the scopes
property, but there is not id_token This leads to an exception.

This change makes the client a little more tolerant of auth
servers by not trying to parse the id_token in this scenario.

#1206